src/requirements.txt
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
decorator==4.3.0
Django==2.0.3
django-debug-toolbar==1.9.1
django-json-rpc==0.7.2
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.9.1
fixture==1.5.11
idna==2.7
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
mock==2.0.0
mysqlclient==1.3.12
parso==0.2.0
pbr==4.3.0
pexpect==4.5.0
pickleshare==0.7.4
pipenv==2018.7.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.3
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
